Subdivide wiring.c
This moves all timer0-related code into a new file wiring_systick.c.
This is a preparation for a next commit, to allow this code to be
included only when linked. The delayMicroseconds function is not
included in this file, since it does not actually use the timer but does
cycle counting (but for consistency, it is moved into its own file
instead).

This commit only moves code and fixes some whitespace in the process, it
does not actually change anything else.
